backstage: bundle o3-components usin tsup
