+++ components/o3-button/src/tsx/group.tsx
@@ -1,4 +1,4 @@
-import type {ButtonGroupProps} from '../types/index.d.ts';
+import type {ButtonGroupProps} from '../types';
 export function ButtonGroup({children = []}: ButtonGroupProps) {
 	return <div className="o3-button-group">{children}</div>;
 }
