--- components/o3-button/src/types/group.d.ts
@@ -1,3 +1,0 @@
-export interface ButtonGroupProps {
-	children: JSX.Element[];
-}
