--- components/o3-button/src/types/index.d.ts
@@ -1,3 +1,0 @@
-export * from './button.d.ts';
-export * from './pagination.d.ts';
-export * from './group.d.ts';
