+++ components/o3-button/src/types/index.ts
@@ -1,0 +1,60 @@
+// Button types
+export interface ButtonProps {
+	label: string;
+	type: 'primary' | 'secondary' | 'ghost';
+	size?: 'small' | '';
+	theme?: 'inverse' | 'mono';
+	icon?:
+		| 'arrow-left'
+		| 'arrow-right'
+		| 'upload'
+		| 'tick'
+		| 'plus'
+		| 'arrow-down'
+		| 'arrow-up'
+		| 'edit-outlined'
+		| 'download'
+		| 'search'
+		| 'refresh'
+		| 'cross';
+	iconOnly?: boolean;
+	visuallyHideDisabled?: boolean;
+	attributes?: {
+		[attribute: string]: string | boolean;
+	};
+	onClick?: Function;
+}
+
+export interface LinkButtonProps extends ButtonProps {
+	href: string;
+}
+
+// Pagination types
+export type ButtonPaginationProps = Pick<ButtonProps, 'theme'> & {
+	previousPager: ButtonPaginationPager;
+	pages: ButtonPaginationItem[];
+	nextPager: ButtonPaginationPager;
+};
+export interface ButtonPaginationItem {
+	href?: string;
+	current: boolean;
+	number: number;
+	onClick?: any;
+}
+export interface ButtonPaginationPager {
+	label: string;
+	href?: string;
+	onClick?: any;
+}
+
+export type Ellipsis = Pick<ButtonProps, 'theme'> & {
+	attributes?: {
+		[attribute: string]: string | boolean;
+	};
+};
+
+
+// ButtonGroup types
+export interface ButtonGroupProps {
+	children: JSX.Element[];
+}
