+++ components/o3-tooltip/tsconfig.json
@@ -1,6 +1,6 @@
 {
 	"extends": "../../tsconfig.json",
 	"include": [
-		"./src/tsx"
+		"./src/tsx", "./src/ts/index.ts"
 	]
 }
