--- components/o3-tooltip/tsconfig.ts.json
@@ -1,8 +1,0 @@
-{
-	"extends": "../../tsconfig.json",
-	"compilerOptions": {
-		"outDir": "./build/js",
-		"lib": ["es2021", "dom"],
-	},
-	"files": ["./src/ts/index.ts"]
-}
