--- components/o3-tooltip/tsconfig.tsx.json
@@ -1,10 +1,0 @@
-{
-	"extends": "../../tsconfig.json",
-	"compilerOptions": {
-		"outDir": "./build/jsx",
-		"declarationDir": "./build/types",
-		"declaration": true,
-		"declarationMap": true
-	},
-	"files": ["./src/tsx/index.ts"]
-}
