+++ tsconfig.json
@@ -5,9 +5,9 @@
 		"jsx": "react-jsx",
 		"esModuleInterop": true,
 		"skipLibCheck": true,
-		"target": "ES2021",
+		"target": "ES2019",
 		"module": "ES2020",
-		"moduleResolution": "bundler",
+		"moduleResolution": "Bundler",
 	},
 	"exclude": ["node_modules"]
 }
